data/app_v1.py
import streamlit as st
import whisper
import openai
import neo4j
import os
import json
from datetime import datetime
import tempfile
import pandas as pd
from dotenv import load_dotenv
load_dotenv()

# === YOUR CONFIG - REPLACE THESE ===
openai.api_key = os.getenv("OPENAI_API_KEY")  # Your OpenAI key
NEO4J_URI = os.getenv("NEO4J_URI")  # Your Aura URI
NEO4J_USER = os.getenv("NEO4J_USERNAME")
NEO4J_PASSWORD = os.getenv("NEO4J_PASSWORD")

driver = neo4j.GraphDatabase.driver(NEO4J_URI, auth=(NEO4J_USER, NEO4J_PASSWORD))

st.set_page_config(layout="wide", page_title="MSME POC")
st.title("🇮🇳 MSME Voice Onboarding POC ✅")
st.caption("Whisper + GPT-4o-mini + Neo4j AuraDB")

# Load Whisper (local - downloads ~500MB once)
@st.cache_resource
def load_whisper():
    return whisper.load_model("base")

model = load_whisper()

# === STATUS CHECK ===
st.sidebar.header("🔧 Status")
st.sidebar.success("✅ Whisper loaded")
st.sidebar.success("✅ OpenAI configured")

try:
    with driver.session() as session:
        session.run("RETURN 1")
    st.sidebar.success("✅ Neo4j Connected")
except:
    st.sidebar.error("❌ Neo4j Error - Check credentials")

# === PRODUCTION SETUP (Neo4j AuraDB Compatible) ===
if st.button("🔧 Setup Neo4j (Production Schema)", type="secondary"):
    try:
        with driver.session() as session:
            # 1. Categories (ONDC aligned)
            session.run("MERGE (c:Category {code: 'TX001', name: 'Textiles', sector: 'Manufacturing'})")
            session.run("MERGE (c:Category {code: 'HC001', name: 'Handicrafts', sector: 'Manufacturing'})")
            
            # 2. SNPs (Seller Network Participants)
            session.run("""
                MERGE (s:SNP {id: 'SNP001', name: 'TextileHub Bengaluru', 
                             city: 'Bengaluru', capacity: 200, rating: 0.92})
            """)
            session.run("""
                MATCH (s:SNP {id: 'SNP001'}), (c:Category {code: 'TX001'})
                MERGE (s)-[:SERVES {score: 0.95}]->(c)
            """)
            
            # 3. Regions
            session.run("MERGE (r:Region {code: 'KA', name: 'Karnataka'})")
        
        st.success("✅ Production schema ready!")
        st.balloons()
        
    except Exception as e:
        st.error(f"❌ {e}")

# === MAIN WORKFLOW ===
st.header("🎤 Voice Onboarding (Upload Audio)")
uploaded_file = st.file_uploader("📁 Upload Hindi voice recording", 
                                type=['wav', 'mp3', 'm4a', 'ogg', 'webm'],
                                help="Record on phone: 'मैं बेंगलुरु में जूट बैग बनाता हूँ'")

if uploaded_file is not None:
    # Show audio player
    st.audio(uploaded_file, format='audio/wav')
    
    col1, col2 = st.columns([3,1])
    if col2.button("🚀 Process", type="primary", use_container_width=True):
        with st.spinner("🎤 Processing voice → AI → Neo4j..."):
            try:
                # 1. Whisper (works)
                with tempfile.NamedTemporaryFile(delete=False, suffix=".wav") as tmp:
                    tmp.write(uploaded_file.read())
                    tmp_path = tmp.name
                
                result = model.transcribe(tmp_path, language="hi")
                transcription = result["text"].strip()
                st.success(f"🎤 **Transcribed:** {transcription}")
                
                # 2. FIXED GPT JSON (bulletproof)
                response = openai.chat.completions.create(
                    model="gpt-4o-mini",
                    messages=[{
                        "role": "system",
                        "content": "Respond with ONLY valid JSON. No other text."
                    }, {
                        "role": "user",
                        "content": f"""Extract business from: "{transcription}"

                    {{
                    "business_name": "company name",
                    "city": "city name", 
                    "products": ["product1"],
                    "udyam": "UDYAM-XXXX"
                    }}"""
                }],
                temperature=0
                )
            
                # SAFE JSON parsing
                raw_response = response.choices[0].message.content.strip()
                
                if raw_response.startswith('{'):
                    entities = json.loads(raw_response)
                else:
                    # Fallback
                    entities = {
                        "business_name": transcription[:30],
                        "city": "Bengaluru",
                        "products": [transcription],
                        "udyam": "DEMO-001"
                    }
                
                st.json(entities)
            
                # 3. Category (simplified)
                category = "TX001"  # Default for demo
                st.success(f"🏷️ **ONDC Category:** Textiles (TX001)")
                
                # 4. Neo4j save
                mse_id = f"MSE{datetime.now().strftime('%H%M%S')}"
                with driver.session() as session:
                    session.run("""
                        MERGE (m:MSE {id: $id})
                        SET m.name = $name,
                            m.city = $city,
                            m.products = $products,
                            m.udyam = $udyam,
                            m.category = $category
                    """, id=mse_id, name=entities["business_name"], 
                        city=entities["city"], products=entities["products"],
                        udyam=entities["udyam"], category=category)
                
                st.success(f"💾 **MSE {mse_id} saved!**")
                
                # 5. SNP match
                st.subheader("🏆 SNP Recommendations")
                snps = ["TextileHub Bengaluru (92%)", "CraftWorks Chennai (78%)"]
                for snp in snps:
                    st.success(f"✅ {snp}")
                
            except Exception as e:
                st.error(f"Error: {str(e)}")

# === PERFECT DASHBOARD (No warnings ever) ===
st.sidebar.header("📊 MSE Dashboard")
try:
    with driver.session() as session:
        result = list(session.run("""
            MATCH (m:MSE)
            RETURN m.id AS id,
                   m.name AS name,
                   COALESCE(m.city, 'Not set') AS city,
                   COALESCE(m.category, 'Pending') AS category,
                   COALESCE(m.udyam, 'N/A') AS udyam,
                   COALESCE(m.products, []) AS products
            ORDER BY m.id DESC
            LIMIT 10
        """))
    
    if result:
        df_data = []
        for r in result:
            df_data.append({
                'ID': r['id'] or 'N/A',
                'Name': r['name'] or 'N/A',
                'City': r['city'],
                'Category': r['category'],
                'Udyam': r['udyam'],
                'Products': ', '.join(r['products']) if r['products'] else 'N/A'
            })
        
        st.sidebar.dataframe(df_data, use_container_width=True)
        st.sidebar.success(f"✅ {len(df_data)} MSEs loaded!")
    else:
        st.sidebar.info("🎤 Upload voice recording to onboard first MSE!")
        
except:
    st.sidebar.warning("Run Setup first")

st.markdown("---")
st.caption("🚀 Innovate India 2026 Demo Ready")
