requirements.txt
streamlit==1.38.0
openai==1.51.2
neo4j==5.24.0
whisper-openai==1.0.0
python-dotenv
transformers==4.46.2
httpx==0.27.2
